# contracts/openapi-slice2.yaml
openapi: 3.1.0
info:
  title: DPP SaaS – SLICE-2 (Organisation & Collaboration)
  description: API-IDs 022–040. Capabilities CAP-006, CAP-007, CAP-008, CAP-009.
  version: 1.0.0
  x-parity-spec: PARITY_SPEC.md
servers:
  - url: /api
    description: API Base
security:
  - sessionCookie: []
paths:
...
  /app/organization/company-details:
    get:
      operationId: API-032-get-company-details
      summary: Firmendaten abrufen
      responses:
        '200': { description: OK }
        '401': { $ref: '#/components/responses/Error401' }
        '500': { $ref: '#/components/responses/Error500' }
    put:
      operationId: API-033-put-company-details
      summary: Firmendaten aktualisieren
      requestBody:
        content:
          application/json:
            schema: { type: object }
      responses:
        '200': { description: OK }
        '401': { $ref: '#/components/responses/Error401' }
        '500': { $ref: '#/components/responses/Error500' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: authjs.session-token
  schemas:
    PlatformError:
      type: object
      required: [error]
      description: Platform-standard error (exakt nach platform constants)
      properties:
        error: { type: string }
      additionalProperties: true
...
  responses:
...
    Error401:
      description: Nicht autorisiert
      content:
        application/json:
          schema: { $ref: '#/components/schemas/PlatformError' }
...
    Error500:
      description: Interner Fehler
      content:
        application/json:
          schema: { $ref: '#/components/schemas/PlatformError' }